Launch alignment: /sponsor server-renders, footer gains /news, /contact routes by audience
/sponsor was a single 'use client' page whose server HTML was just
'Loading…' — the narrative shell (hero, journey, gallery, CTA) now renders
server-side; the cart form and newsletter capture stay client, still
pricing only from /api/products.

Footer: /news added to About; duplicate 'Back the work' dropped (the
/partner-vs-/invest rename waits for #214).

Contact: quick links become audience doors — shop, sponsor, communities,
press, backing the work — and the dead /about link goes straight to the
real routes.
