+++ v2/src/app/sponsor/palette.ts
@@ -1,0 +1,6 @@
+// Palette — matches the Canberra hero so the cross-page flow feels like one site.
+// Shared by the server page and the client form/newsletter components.
+export const CREAM = '#FDF8F3';
+export const RUST = '#C45C3E';
+export const CHARCOAL = '#2E2E2E';
+export const SAGE = '#8B9D77';
